backend-app/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
Flask==3.1.3
flask-cors==6.0.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.33.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.8
